requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
argilla==1.12.0
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
chromadb==0.3.23
click==8.1.3
clickhouse-connect==0.6.4
colorama==0.4.6
colorclass==2.2.2
comm==0.1.4
commonmark==0.9.1
compressed-rtf==1.0.6
contourpy==1.0.7
cryptography==41.0.1
curlify==2.2.1
cycler==0.11.0
dataclasses-json==0.5.9
debugpy==1.6.7.post1
decorator==5.1.1
Deprecated==1.2.14
duckdb==0.8.1
easygui==0.98.3
ebcdic==1.1.1
einops==0.6.1
et-xmlfile==1.1.0
executing==1.2.0
extract-msg==0.41.1
fastapi==0.99.0
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.12.0
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpt4all==0.3.4
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.54.0
h11==0.14.0
h5py==3.8.0
hnswlib==0.7.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.15.1
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
imagezmq==1.1.1
IMAPClient==2.3.1
imbalanced-learn==0.11.0
ipykernel==6.25.1
ipython==8.14.0
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
keras==2.13.1
kiwisolver==1.4.4
langchain==0.0.197
langchainplus-sdk==0.0.19
lark-parser==0.12.0
libclang==16.0.0
llama-cpp-python==0.1.50
llvmlite==0.40.1
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
ml-dtypes==0.1.0
mlxtend==0.22.0
mne==1.3.1
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
msoffcrypto-tool==5.0.1
mss==9.0.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
numba==0.57.1
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
olefile==0.46
oletools==0.60.1
openapi-schema-pydantic==1.2.4
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==1.5.3
pandoc==2.3
parso==0.8.3
pcodedmp==1.2.6
pdfminer.six==20221105
pickleshare==0.7.5
Pillow==9.5.0
pip==23.2.1
platformdirs==3.5.0
plotly==5.15.0
plumbum==1.8.2
ply==3.11
pooch==1.7.0
posthog==3.0.1
prompt-toolkit==3.0.39
protobuf==4.22.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.10
PyDispatcher==2.0.7
pydot==1.4.2
pyEDFlib==0.1.32
pygame==2.4.0
Pygments==2.15.1
PyMuPDF==1.22.3
pynndescent==0.5.10
pypandoc==1.11
pyparsing==2.4.7
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyriemann==0.4
PySide6==6.5.0
PySide6-Addons==6.5.0
PySide6-Essentials==6.5.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0
pyzmq==25.1.1
red-black-tree-mod==1.20
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.0.1
rsa==4.9
RTFDE==0.0.2
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.2
screeninfo==0.8.1
seaborn==0.12.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==68.1.2
shiboken6==6.5.0
six==1.16.0
sklearn==0.0.post7
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.17
stack-data==0.6.2
starlette==0.27.0
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
torchinfo==1.7.2
torchvision==0.15.2
tornado==6.3.3
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
typer==0.7.0
typing_extensions==4.5.0
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==4.2
umap-learn==0.5.3
unstructured==0.6.6
urllib3==2.0.2
uvicorn==0.22.0
watchfiles==0.19.0
wcwidth==0.2.6
websocket-client==1.5.1
websockets==11.0.2
Werkzeug==2.3.3
wheel==0.41.1
win-unicode-console==0.5
wrapt==1.14.1
XlsxWriter==3.1.2
yarl==1.9.2
zstandard==0.21.0


#pip install -r ./requirements.txt
